refactor: :recycle: move showable + messagingdelegate
